move getAuthor from Records util to core Message class
