+++ src/interfaces/records-write.ts
@@ -251,7 +251,7 @@ export class RecordsWrite implements MessageInterface<RecordsWriteMessage> {
     this._message = message;
 
     if (message.authorization !== undefined) {
-      this._author = Records.getAuthor(message as RecordsWriteMessage);
+      this._author = Message.getAuthor(message as RecordsWriteMessage);
 
       this._signaturePayload = Jws.decodePlainObjectPayload(message.authorization.signature);
 
@@ -831,7 +831,7 @@ export class RecordsWrite implements MessageInterface<RecordsWriteMessage> {
     }
 
     const recordsWriteMessage = message as RecordsWriteMessage;
-    const author = Records.getAuthor(recordsWriteMessage);
+    const author = Message.getAuthor(recordsWriteMessage);
     const entryId = await RecordsWrite.getEntryId(author, recordsWriteMessage.descriptor);
     return (entryId === recordsWriteMessage.recordId);
   }
