+++ tests/core/message.spec.ts
@@ -1,12 +1,70 @@
+import type { PermissionScope } from '../../src/types/permission-types.js';
 import type { RecordsQueryReplyEntry } from '../../src/types/records-types.js';
 
 import { expect } from 'chai';
+import { Jws } from '../../src/utils/jws.js';
 import { Message } from '../../src/core/message.js';
-import { RecordsRead } from '../../src/index.js';
+import { PermissionsProtocol } from '../../src/protocols/permissions.js';
+import { RecordsRead } from '../../src/interfaces/records-read.js';
+import { RecordsWrite } from '../../src/interfaces/records-write.js';
 import { TestDataGenerator } from '../utils/test-data-generator.js';
 import { Time } from '../../src/utils/time.js';
+import { DwnInterfaceName, DwnMethodName } from '../../src/enums/dwn-interface-method.js';
 
 describe('Message', () => {
+  describe('getAuthor()', () => {
+    it('should return the author of RecordsWrite, RecordsDelete types', async () => {
+      const bob = await TestDataGenerator.generatePersona();
+
+      // create a record message
+      const { message: recordsWriteMessage } = await TestDataGenerator.generateRecordsWrite({ author: bob });
+      const recordsWriteAuthor = Message.getAuthor(recordsWriteMessage);
+      expect(recordsWriteAuthor).to.equal(bob.did);
+
+      // create a delete message
+      const { message: recordsDeleteMessage } = await TestDataGenerator.generateRecordsDelete({ author: bob });
+      const recordsDeleteAuthor = Message.getAuthor(recordsDeleteMessage);
+      expect(recordsDeleteAuthor).to.equal(bob.did);
+    });
+
+    it('should get the author of a delegated message', async () => {
+      const alice = await TestDataGenerator.generatePersona();
+      const deviceX = await TestDataGenerator.generatePersona();
+
+      // create a delegation scope from alice to deviceX for writing records in a protocol
+      const scope:PermissionScope = {
+        interface : DwnInterfaceName.Records,
+        method    : DwnMethodName.Write,
+        protocol  : 'https://example.com/protocol/test',
+      };
+
+      // create the delegated grant message
+      const bobGrant = await PermissionsProtocol.createGrant({
+        delegated   : true,
+        dateExpires : Time.createOffsetTimestamp({ seconds: 100 }),
+        grantedTo   : deviceX.did,
+        scope       : scope,
+        signer      : Jws.createSigner(alice)
+      });
+
+      // create a record message using the grant
+      const writeData = TestDataGenerator.randomBytes(32);
+
+      const { message } = await RecordsWrite.create({
+        signer         : Jws.createSigner(deviceX),
+        delegatedGrant : bobGrant.dataEncodedMessage,
+        protocol       : 'https://example.com/protocol/test',
+        protocolPath   : 'test/path',
+        dataFormat     : 'application/json',
+        data           : writeData,
+      });
+
+      // expect message author to be alice
+      const author = Message.getAuthor(message);
+      expect(author).to.equal(alice.did);
+    });
+  });
+
   describe('getSigner()', () => {
     it('should return `undefined` if message is not signed', async () => {
       const recordsRead = await RecordsRead.create({
